Add some basics (#3)
* Add simple tileset for testing.

Going to be using these sprites for the tileset until we get some actual sprites drawn.

* Set project width/height to 240x180.

Just setting a default width/height for now. Also set the stretch mode to 2D.

* Add "empty" section to tileset.

I'm not sure how we can "hide" the upper parts of objects, so for now i'm just using this empty tile.

* Create Maps directory + Outside tilemap.

We can keep our (scenes with tile)maps in the Maps directory. I also created a simple tilemap where the player can start the game.

* Added a Dogtective sprite.

Behold... Dogtective!

* Create Dogtective scene.

Added Dogtective to the Outside scene and connected Dogtective to the camera.

* Add tile by tile movement.

Dogtective moves tile by tile and collides with the environment.

* Restore window size.

Window size was removed when merging from master.

* Change Dogtective movement to pixel by pixel and add camera scrolling.

* Create Camera2D scene with defaults.

Easily share the Camera2D with sane defaults across all maps.

* Change from RectangleShape2D to CapsuleShape2D.

Before, i was using a rectangle shape to handle the tile by tile movment. Now that we're going JRPG style we can have more detailed collisions.

* Transitions between maps.

Added an "Actions" folder and merged the 'scenes' folder into the 'Maps' folder.

Created an "action" called "MapChange" that you can place on the map. When the player touches that area they will be transported to the map set in the 'map_name' export variable at the coordinates in specified in the 'coordinates' export variable.

* Add thought bubbles.

Thought Bubbles show up for a short period of time and don't pause the game. It just provides some interaction with the environment to add some humor or push the story forward.

To create a thought bubble, you need to add an Action/ThoughtBubble node to the scene and put the text you'd like drawn in the export variable.

Co-authored-by: chickendude <pokemaster103@gmail.com>